change value to float

+++ database/migrations/2025_11_01_120948_create_nutrientables_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('nutrientables', function (Blueprint $table) {
             $table->foreignId('nutrition_id')->constrained()->cascadeOnDelete();
             $table->morphs('nutrientable');
-            $table->unsignedInteger('value')->nullable();
+            $table->float('value', 8)->nullable();
             $table->unsignedInteger('percentage')->nullable();
         });
     }
